Added increment to LinkStrength every time a Mail is read from a Link

+++ link.go
@@ -2,8 +2,7 @@ package orgnetsim
 
 //A Link between Agents in the Network
 type Link struct {
-	Agent1ID string  `json:"agent1Id"`
-	Agent2ID string  `json:"agent2Id"`
-	Strength float64 `json:"strength"`
-	
+	Agent1ID string `json:"agent1Id"`
+	Agent2ID string `json:"agent2Id"`
+	Strength int    `json:"strength"`
 }
